ElectroStock/requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
cachecontrol==0.13.1
cachetools==5.3.1
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==40.0.2
DateTime==5.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-adminlte2==0.4.3
django-allauth==0.54.0
django-appconf==1.0.5
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-bootstrap4==23.2
django-compat==1.0.15
django-cors-headers==3.14.0
django-cron==0.6.0
django-datetimepicker==3.14
django-dbbackup==4.0.2
django-fontawesome-5==1.0.18
django-import-export==3.2.0
django-jazzmin==2.6.0
django-select2==8.1.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
et-xmlfile==1.1.0
firebase==4.0.1
firebase-admin==6.2.0
fontawesome-free==5.15.4
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupPy==1.14
MarkupSafe==2.1.2
msgpack==1.0.5
Naked==0.1.32
numpy==1.24.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.5.0
Pillow-SIMD==9.0.0.post1
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.4.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
validate-email-address==1
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
colorama==0.4.6
gunicorn==21.2.0
packaging==23.2
